requirements.txt
flask
opencv-python-headless
ultralytics
norfair
psycopg2-binary
numpy
requests
python-dotenv